Django 1.8 Compatibility - Set Custom Address/Port
Added the ability to set a custom address/port in Django 1.8. In Django 1.7 and previous versions this could be set as an argument:

./manage.py runsslserver 0:8080 --certificate ...

Now you must pass that argument as an option with --addrport (In django's new runserver command), but runsslserver doesn't recognize that option as valid which causes it to send a false negative error message. This change will fix that.